fix(hr): 修 CodeRabbit 在 PR #35 抓到的三處（逐條查證後皆成立）
## 1. 🟠 部分更新繞過時間順序檢查

`correct_attendance` 原本只驗 request 帶來的兩個欄位。更正請求可以只帶一邊——
只送 `clock_in_time=18:00`、而既有紀錄 `clock_out_time=17:00` 時，那個 tuple match
根本不成立（另一邊是 None），負區間被寫進 DB，`compute_regular_hours` 又把它算成
0.0：資料庫留下一筆下班早於上班、工時 0 的紀錄，且沒有任何錯誤訊息。

改為抽出 `validate_final_time_order`，驗的是 `final_in` / `final_out`（合併既有值
之後的最終時間），並在 `backfill_attendance` 走同一個函式避免兩條路徑日後漂移。

## 2. 🟡 Excel 合計列漏兩欄

畫面上的合計列有 incomplete_days / corrected_days，Excel 只寫到第 4 欄——
看報表的人會以為那兩欄沒有值。合計邏輯抽成 `MonthlyReportTotals` 讓它可被測試，
合計列補齊每一個數值欄。

## 3. 🟡 Decimal 序列化成字串，前端宣告 number

`rust_decimal` 只開 `serde` feature 時輸出 JSON **字串**，而前端
`MonthlyAttendanceSummary` 宣告 `number`。字串進 `useTableSort` 的 `compareValues`
會走字串比較——`"9.5"` 排在 `"168.5"` 後面（實查 `useTableSort.ts:31`：
只在 `typeof === 'number'` 才做數值比較）。

不改 Cargo.toml 加 `serde-float`——那會全域改變所有 Decimal 欄位的序列化，
等於動到既有 API contract。改在 SQL 端 `::float8` + 欄位型別 `f64`，與同檔
`AttendanceStat.overtime_hours` 的既有寫法一致。來源是 `numeric(5,2)`，
月度加總的精度遠在 f64 安全範圍內。前端 `formatHours` 的 parseFloat 防禦分支
隨之成為死碼，一併移除。

## 驗證

- `cargo test --lib` 36 passed（31 → 36，新增 5 條）
- **mutation 驗過**：把兩處修正各自改回壞寫法 → `partial_correction_with_inverted_
  merged_range_is_rejected`、`equal_in_and_out_is_rejected`、
  `totals_cover_every_numeric_column` 三條如期轉紅；還原後 36 全綠
- `cargo clippy --all-targets -- -D warnings -W clippy::unwrap_used` exit 0
- `cargo fmt --check` exit 0
- `vitest run --coverage` exit 0（lines 9.98% / functions 6.47%，門檻未動）
- `tsc --noEmit` / `eslint --max-warnings=0` exit 0

+++ backend/src/models/hr.rs
@@ -132,8 +132,15 @@ pub struct MonthlyAttendanceSummary {
     pub user_email: String,
     /// 當月有「上班打卡時間」的天數
     pub work_days: i64,
-    pub total_regular_hours: Decimal,
-    pub total_overtime_hours: Decimal,
+    /// ⚠️ 型別是 `f64` 不是 `Decimal`（CodeRabbit PR #35 指出）：
+    /// `rust_decimal` 只開 `serde` feature 時會把 Decimal 序列化成**字串**，
+    /// 而前端 `MonthlyAttendanceSummary` 宣告的是 `number`。字串進到
+    /// `useTableSort` 的 `compareValues` 會走字串比較——`"9.5"` 排在 `"168.5"` 後面。
+    /// 改 Cargo.toml 加 `serde-float` 會全域改變所有 Decimal 欄位的序列化（等於動到
+    /// 既有 API contract），故改在 SQL 端 `::float8`，與同檔 `AttendanceStat.overtime_hours`
+    /// 的既有寫法一致。來源欄位是 `numeric(5,2)`，月度加總的精度遠在 f64 安全範圍內。
+    pub total_regular_hours: f64,
+    pub total_overtime_hours: f64,
     /// 上下班卡只有一邊的天數——這些日子的工時不完整，是補卡的候選清單
     pub incomplete_days: i64,
     /// 當月經補登／更正的天數（`is_corrected`）
